requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
black==23.12.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
